+++ src/annotations.ts
@@ -35,116 +35,162 @@ Portions of this software are derived from [ghcid](https://github.com/ndmitchell
 
 import * as vscode from 'vscode'
 
-import { Config, Elem, alloglot } from "./config";
+import { Annotation, AnnotationsConfig, LanguageConfig, alloglot } from "./config";
 
-export function makeAnnotations(lang: Elem<Config['languages']>): vscode.Disposable {
-  const files = lang.annotationsFiles
-  if (files.length === 0) return { dispose: () => { } }
+export function makeAnnotations(config: LanguageConfig): vscode.Disposable {
+  const { languageId, annotations } = config
+  if (!languageId || !annotations || annotations.length === 0) return vscode.Disposable.from()
 
   const diagnostics = vscode.languages.createDiagnosticCollection(alloglot.collections.annotations)
-  const watchers = files.map(watchAnnotationsFile(diagnostics)).flat()
+  const watchers: Array<vscode.Disposable> = annotations.map(file => watchAnnotationsFile(diagnostics, file))
 
-  const actions = vscode.languages.registerCodeActionsProvider(
-    lang.languageId,
-    { provideCodeActions: (document, range, context) => context.diagnostics.map(asQuickFixes).flat() },
+  const quickFixes = vscode.languages.registerCodeActionsProvider(
+    languageId,
+    { provideCodeActions: (document, range, context) => context.diagnostics.map(utils.asQuickFixes).flat() },
     { providedCodeActionKinds: [vscode.CodeActionKind.QuickFix] }
   )
 
   return vscode.Disposable.from(
-    actions,
+    quickFixes,
     ...watchers,
     diagnostics
   )
 }
 
-function watchAnnotationsFile(diagnostics: vscode.DiagnosticCollection): (file: string) => vscode.Disposable {
-  return file => {
-    const watchers = vscode.workspace.workspaceFolders?.map(ws => {
-      const pattern = new vscode.RelativePattern(ws, file)
-      const watcher = vscode.workspace.createFileSystemWatcher(pattern, false, false, false)
+function watchAnnotationsFile(diagnostics: vscode.DiagnosticCollection, cfg: AnnotationsConfig): vscode.Disposable {
+  const messagePath = utils.path<string>(cfg.mapping.message)
+  const filePath = utils.path<string>(cfg.mapping.file)
+  const startLinePath = utils.path<number>(cfg.mapping.startLine)
+  const startColumnPath = utils.path<number>(cfg.mapping.startColumn)
+  const endLinePath = utils.path<number>(cfg.mapping.endLine)
+  const endColumnPath = utils.path<number>(cfg.mapping.endColumn)
+  const sourcePath = utils.path<string>(cfg.mapping.source)
+  const severityPath = utils.path<string>(cfg.mapping.severity)
+  const replacementsPath = utils.path<Array<string>>(cfg.mapping.replacements)
+  const referenceCodePath = utils.path<string | number>(cfg.mapping.referenceCode)
 
-      watcher.onDidChange(addAnnotations(diagnostics))
-      watcher.onDidCreate(addAnnotations(diagnostics))
-      watcher.onDidDelete(uri => diagnostics.delete(uri))
+  function marshalAnnotation(json: any): Annotation | undefined {
+    const message = messagePath(json)
+    const file = filePath(json)
+    if (!message || !file) return
 
-      return watcher
-    }) || []
+    const startLine = startLinePath(json) || 0
+    const startColumn = startColumnPath(json) || 0
+    const endLine = endLinePath(json) || startLine
+    const endColumn = endColumnPath(json) || startColumn
 
-    return vscode.Disposable.from(...watchers)
+    return {
+      message, file, startLine, startColumn, endLine, endColumn,
+      source: sourcePath(json) || `${cfg.file}`,
+      severity: utils.parseSeverity(severityPath(json)),
+      replacements: replacementsPath(json) || [],
+      referenceCode: referenceCodePath(json)?.toString(),
+    }
   }
-}
 
-type Annotation = {
-  source: string
-  severity: 'error' | 'warning' | 'info' | 'hint'
-  file: string
-  startLine: number
-  startColumn: number
-  endLine: number
-  endColumn: number
-  message: string
-  replacements: Array<string>
-  referenceCode?: string
-}
+  function readAnnotations(bytes: Uint8Array): Array<Annotation> {
+    const contents = Buffer.from(bytes).toString('utf8')
+    const jsons: Array<any> = cfg.format === 'jsonl'
+      ? contents.split('\n').map(line => JSON.parse(line))
+      : JSON.parse(contents)
+    const annotations = jsons.map(marshalAnnotation).filter(x => x) as Array<Annotation>
+    return annotations
+  }
 
-function addAnnotations(diagnostics: vscode.DiagnosticCollection): (uri: vscode.Uri) => void {
-  return uri => {
+  function annotationsBySourceFile(annotations: Array<Annotation>): Map<string, Array<Annotation>> {
+    const sorted = new Map<string, Array<Annotation>>()
+    annotations.forEach(annotation => {
+      const annotationsForFile = sorted.get(annotation.file)
+      annotationsForFile
+        ? annotationsForFile.push(annotation)
+        : sorted.set(annotation.file, [annotation])
+    })
+    return sorted
+  }
+
+  function addAnnotations(uri: vscode.Uri): void {
     vscode.workspace.fs.readFile(uri).then(bytes => {
-      const annotations: Array<Annotation> = JSON.parse(Buffer.from(bytes).toString('utf8')) || []
-      const sorted: Map<string, Array<Annotation>> = new Map()
-      annotations.forEach(ann => {
-        const current = sorted.get(ann.file)
-        current
-          ? current.push(ann)
-          : sorted.set(ann.file, [ann])
-      })
-      sorted.forEach((anns, path) => {
-        diagnostics.set(vscode.Uri.file(path), anns.map(asDiagnostic))
+      annotationsBySourceFile(readAnnotations(bytes)).forEach((anns, path) => {
+        diagnostics.set(vscode.Uri.file(path), anns.map(utils.annotationAsDiagnostic))
       })
     })
   }
+
+  const watchers = vscode.workspace.workspaceFolders?.map(ws => {
+    const pattern = new vscode.RelativePattern(ws, cfg.file)
+    const watcher = vscode.workspace.createFileSystemWatcher(pattern, false, false, false)
+
+    watcher.onDidChange(addAnnotations)
+    watcher.onDidCreate(addAnnotations)
+    watcher.onDidDelete(diagnostics.delete)
+
+    return watcher
+  }) || []
+
+  return vscode.Disposable.from(...watchers)
 }
 
-function asDiagnostic(ann: Annotation): vscode.Diagnostic {
-  const range = annotationRange(ann)
-  const message = ann.message
-  const severity = asDiagnosticSeverity(ann.severity)
-  const diag = new vscode.Diagnostic(range, message, severity)
-  diag.source = ann.source
-  diag.relatedInformation = annotationRelatedInformation(ann)
-  diag.code = diag.code
-  return diag
-}
+namespace utils {
+  export function annotationAsDiagnostic(ann: Annotation): vscode.Diagnostic {
+    const range = new vscode.Range(
+      new vscode.Position(ann.startLine - 1, ann.startColumn - 1),
+      new vscode.Position(ann.endLine - 1, ann.endColumn - 1)
+    )
 
-function annotationRelatedInformation(ann: Annotation): Array<vscode.DiagnosticRelatedInformation> {
-  const uri = vscode.Uri.file(ann.file)
-  const range = annotationRange(ann)
-  const location = new vscode.Location(uri, range)
-  return ann.replacements.map(message => new vscode.DiagnosticRelatedInformation(location, message))
-}
+    // we are abusing the relatedInformation field to store replacements
+    // we look them up later when we need to create quick fixes
+    const relatedInformation = ann.replacements.map(replacement => {
+      const uri = vscode.Uri.file(ann.file)
+      const location = new vscode.Location(uri, range)
+      return new vscode.DiagnosticRelatedInformation(location, replacement)
+    })
 
-function annotationRange(ann: Annotation): vscode.Range {
-  const start = new vscode.Position(ann.startLine - 1, ann.startColumn - 1)
-  const end = new vscode.Position(ann.endLine - 1, ann.endColumn - 1)
-  return new vscode.Range(start, end)
-}
-
-function asDiagnosticSeverity(sev: Annotation['severity']): vscode.DiagnosticSeverity {
-  switch (sev) {
-    case 'error': return vscode.DiagnosticSeverity.Error
-    case 'warning': return vscode.DiagnosticSeverity.Warning
-    case 'info': return vscode.DiagnosticSeverity.Information
-    case 'hint': return vscode.DiagnosticSeverity.Hint
+    // i wish they gave an all-args constructor
+    const diagnostic = new vscode.Diagnostic(range, ann.message, asDiagnosticSeverity(ann.severity))
+    diagnostic.source = ann.source
+    diagnostic.relatedInformation = relatedInformation
+    diagnostic.code = ann.referenceCode
+    return diagnostic
   }
-}
 
-function asQuickFixes(diag: vscode.Diagnostic): Array<vscode.CodeAction> {
-  const actions = diag.relatedInformation?.map(info => {
-    const action = new vscode.CodeAction(diag.message, vscode.CodeActionKind.QuickFix)
-    action.diagnostics = [diag]
-    action.edit = new vscode.WorkspaceEdit
-    action.edit.replace(info.location.uri, info.location.range, info.message)
-    return action
-  })
-  return actions || []
+  function asDiagnosticSeverity(sev: Annotation['severity']): vscode.DiagnosticSeverity {
+    switch (sev) {
+      case 'error': return vscode.DiagnosticSeverity.Error
+      case 'warning': return vscode.DiagnosticSeverity.Warning
+      case 'info': return vscode.DiagnosticSeverity.Information
+      case 'hint': return vscode.DiagnosticSeverity.Hint
+    }
+  }
+
+  // this depends on the fact that we're abusing the `relatedInformation` field
+  // see `annotationAsDiagnostic` above
+  export function asQuickFixes(diag: vscode.Diagnostic): Array<vscode.CodeAction> {
+    const actions = diag.relatedInformation?.map(info => {
+      const action = new vscode.CodeAction(diag.message, vscode.CodeActionKind.QuickFix)
+      action.diagnostics = [diag]
+      action.edit = new vscode.WorkspaceEdit
+      action.edit.replace(info.location.uri, info.location.range, info.message)
+      return action
+    })
+    return actions || []
+  }
+
+  export function path<T>(keys: Array<string> | undefined): (json: any) => T | undefined {
+    if (!keys) return () => undefined
+    else return json => {
+      const result = keys.reduce((acc, key) => acc?.[key], json)
+      if (result) return result as T
+      else return
+    }
+  }
+
+  export function parseSeverity(raw: string | undefined): Annotation['severity'] {
+    if (!raw) return 'error'
+    const lower = raw.toLowerCase()
+    if (lower.includes('error')) return 'error'
+    if (lower.includes('warning')) return 'warning'
+    if (lower.includes('info')) return 'info'
+    if (lower.includes('hint')) return 'hint'
+    return 'error'
+  }
 }