+++ src/config.ts
@@ -1,13 +1,99 @@
-export type Elem<as extends Array<unknown>> = as extends Array<infer a> ? a : never
-
+/**
+ * Extension configuration.
+ */
 export type Config = {
-  languages: Array<{
-    languageId: string
-    serverCommand: string
-    formatCommand: string
-    apiSearchUrl: string
-    annotationsFiles: Array<string>
-  }>
+  /**
+   * An array of per-language configurations.
+   */
+  languages: Array<LanguageConfig>
+}
+
+/**
+ * Configuration for an arbitrary language.
+ */
+export type LanguageConfig = {
+  /**
+   * The unique language ID.
+   * You can usually find this in a language's syntax-highlighting extension.
+   */
+  languageId: string
+
+  /**
+   * A command to start the language server.
+   */
+  serverCommand?: string
+
+  /**
+   * A formatter command.
+   * Reads from STDIN and writes to STDOUT.
+   * `${file}` will be interpolated with the path to the file.
+   */
+  formatCommand?: string
+
+  /**
+   * URL to documentation/API search.
+   * `${query}` will be interpolated with the symbol under cursor.
+   */
+  apiSearchUrl?: string
+
+  /**
+   * A list of files to watch for compiler-generated JSON output.
+   */
+  annotations?: Array<AnnotationsConfig>
+}
+
+/**
+ * A file to watch for compiler-generated JSON output, and instructions on how to marshal the JSON objects.
+ */
+export type AnnotationsConfig = {
+  /**
+   * The path to the file to watch.
+   */
+  file: string
+
+  /**
+   * `json` for a top-level array of objects.
+   * `jsonl` for a newline-separated stream of objects.
+   */
+  format: 'json' | 'jsonl'
+
+  /**
+   * Mapping between properties of the JSON objects and properties of `Annotation`.
+   */
+  mapping: AnnotationsMapping
+}
+
+/**
+ * Intermediate representation of compiler-generated JSON output and VS Code diagnostics.
+ */
+export type Annotation = {
+  source: string
+  severity: 'error' | 'warning' | 'info' | 'hint'
+  file: string
+  startLine: number
+  startColumn: number
+  endLine: number
+  endColumn: number
+  message: string
+  replacements: Array<string>
+  referenceCode?: string
+}
+
+/**
+ * Mapping between arbitrary JSON object and properties of `Annotation`.
+ * Each property is an array of strings that will be used as a path into the JSON object.
+ */
+export type AnnotationsMapping = {
+  message: Array<string>
+  file?: Array<string>
+  startLine?: Array<string>
+  startColumn?: Array<string>
+  endLine?: Array<string>
+  endColumn?: Array<string>
+  source?: Array<string>
+  severity?: Array<string>
+  replacements?: Array<string>
+  referenceCode?: Array<string>
 }
 
 export namespace alloglot {